# example/api/openapiv2/hello-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "hello-service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "HelloService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/hello": {
      "get": {
        "operationId": "HelloService_SayHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiHelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HelloService"
        ]
      }
    },
    "/v1/hello:stream": {
      "get": {
        "operationId": "HelloService_SayHelloStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/apiHelloResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of apiHelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HelloService"
        ]
      }
    }
  },
  "definitions": {
    "apiHelloResponse": {
      "type": "object",
      "properties": {
        "reply": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}